create new team before upgrade
